move models to their own package
