+++ db/model/user_activity.go
@@ -1,4 +1,4 @@
-package db
+package model
 
 import (
 	"time"
@@ -14,15 +14,4 @@ type UserActivity struct {
 	VoiceCount                 int
 	LastChannelVoiceAppearence *time.Time
 	LastChannelTextAppearence  *time.Time
-}
-
-//NitroUserChannel represents a voice channel registered to a nitro booster
-type NitroUserChannel struct {
-	gorm.Model
-	Name      string
-	UserID    string
-	ChannelID string
-	Active    bool
-	LastUsed  *time.Time
-	Enabled   bool
 }
