+++ handler/user_activity_handler.go
@@ -6,6 +6,7 @@ import (
 	"github.com/bwmarrin/discordgo"
 	"github.com/rs/zerolog/log"
 	"github.com/the-sanctuary/waddles/db"
+	"github.com/the-sanctuary/waddles/db/model"
 	"github.com/the-sanctuary/waddles/util"
 )
 
@@ -13,9 +14,9 @@ func UserActivityTextChannel(s *discordgo.Session, m *discordgo.MessageCreate) {
 	if m.Author.ID == s.State.User.ID {
 		return
 	}
-	var ua db.UserActivity
+	var ua model.UserActivity
 
-	r := db.CurrentWadlDB().DB.Where(&db.UserActivity{UserID: m.Author.ID}).FirstOrCreate(&ua)
+	r := db.CurrentWadlDB().DB.Where(&model.UserActivity{UserID: m.Author.ID}).FirstOrCreate(&ua)
 	util.DebugError(r.Error)
 
 	now := time.Now()
@@ -30,11 +31,11 @@ func UserActivityVoiceChannel(s *discordgo.Session, vsu *discordgo.VoiceStateUpd
 	if vsu.UserID == s.State.User.ID {
 		return
 	}
-	var ua db.UserActivity
+	var ua model.UserActivity
 
 	log.Debug().Msgf("Voice session id: %s", vsu.SessionID)
 
-	r := db.CurrentWadlDB().DB.Where(&db.UserActivity{UserID: vsu.UserID}).FirstOrCreate(&ua)
+	r := db.CurrentWadlDB().DB.Where(&model.UserActivity{UserID: vsu.UserID}).FirstOrCreate(&ua)
 	util.DebugError(r.Error)
 
 	now := time.Now()
